Move LocationClient to new module
