+++ collect_app/src/main/java/org/odk/collect/android/location/client/GoogleFusedLocationClient.java
@@ -20,6 +20,7 @@ import com.google.android.gms.location.LocationRequest;
 import com.google.android.gms.location.LocationServices;
 
 import org.odk.collect.android.utilities.GeoUtils;
+import org.odk.collect.location.LocationClient;
 
 import timber.log.Timber;
