+++ collect_app/src/test/java/org/odk/collect/android/location/client/AndroidLocationClientTest.java
@@ -26,10 +26,10 @@ import static org.junit.Assert.assertSame;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.odk.collect.android.location.client.LocationClient.Priority.PRIORITY_BALANCED_POWER_ACCURACY;
-import static org.odk.collect.android.location.client.LocationClient.Priority.PRIORITY_HIGH_ACCURACY;
-import static org.odk.collect.android.location.client.LocationClient.Priority.PRIORITY_LOW_POWER;
-import static org.odk.collect.android.location.client.LocationClient.Priority.PRIORITY_NO_POWER;
+import static org.odk.collect.location.LocationClient.Priority.PRIORITY_BALANCED_POWER_ACCURACY;
+import static org.odk.collect.location.LocationClient.Priority.PRIORITY_HIGH_ACCURACY;
+import static org.odk.collect.location.LocationClient.Priority.PRIORITY_LOW_POWER;
+import static org.odk.collect.location.LocationClient.Priority.PRIORITY_NO_POWER;
 
 @RunWith(AndroidJUnit4.class)
 public class AndroidLocationClientTest {
