+++ collect_app/src/test/java/org/odk/collect/android/location/client/FakeLocationClient.java
@@ -2,11 +2,13 @@ package org.odk.collect.android.location.client;
 
 import android.location.Location;
 
+import androidx.annotation.Nullable;
+
 import com.google.android.gms.location.LocationListener;
 
-import java.lang.ref.WeakReference;
+import org.odk.collect.location.LocationClient;
 
-import androidx.annotation.Nullable;
+import java.lang.ref.WeakReference;
 
 public class FakeLocationClient implements LocationClient {
     private boolean failOnStart;
