+++ collect_app/src/test/java/org/odk/collect/android/location/client/LocationClientProviderTest.java
@@ -21,6 +21,7 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.odk.collect.android.utilities.PlayServicesChecker;
+import org.odk.collect.location.LocationClient;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.instanceOf;
