+++ location/src/main/java/org/odk/collect/location/LocationClient.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.location.client;
+package org.odk.collect.location;
 
 import android.location.Location;
