+++ settings.gradle
@@ -13,3 +13,4 @@ include ':collect_app'
 include ':nbistubs'
 include ':material'
 include ':fragmentstest'
+include ':location'
